third commit

+++ controllers/user.controller.js
@@ -5,7 +5,9 @@ const Mongo = require("../models/user");
 const config = require("config");
 
 const myJwt = require("../services/jwt-service");
-const user = require("../models/user");
+
+const uuid = require("uuid");
+const mailService = require("../services/mail-service");
 
 // const generateAccessToken = (id, name, email) => {
 //   const payload = {
@@ -20,16 +22,46 @@ const add = async (req, res) => {
   try {
     //Add
     const { name, email, password } = req.body;
-    if (name == "" || email == "" || password == "") {
-      return res.status(400).send({ message: "Ma'lumotlarni to'liq yuboring" });
-    }
+
+    const user_activation_link = uuid.v4();
     const neww = await Mongo({
       name,
       email,
       password,
+      user_activation_link,
     });
     await neww.save();
-    res.status(200).send({ message: "Qo'shildi" });
+
+    await mailService.sendActivationMail(
+      email,
+      `${config.get("api_url")}/api/users/activate/${user_activation_link}`
+    );
+
+    const payload = {
+      id: neww._id,
+      name: neww.name,
+      email: neww.email,
+      user_is_active: neww.user_is_active,
+    };
+    const tokens = myJwt.generateTokens(payload);
+    neww.user_token = tokens.refreshToken;
+    await neww.save();
+
+    res.cookie("refreshToken", tokens.refreshToken, {
+      maxAge: config.get("refresh_ms"),
+      httpOnly: true,
+    });
+
+    // try {
+    //   setTimeout(function () {
+    //     var err = new Error("Hello");
+    //     throw err;
+    //   }, 1000);
+    // } catch (e) {
+    //   console.log(e.message);
+    //   res.status(404).send({ message: e.message });
+    // }
+    res.status(200).send({ ...tokens, user: payload });
   } catch (error) {
     errorHandler(res, error);
   }
@@ -105,6 +137,10 @@ const deleteOne = async (req, res) => {
   }
   try {
     const myId = req.params.id;
+    if (myId !== req.user.id) {
+      return res.status(401).send({ message: "Sizda bunday huquq yo'q" });
+    }
+
     await Mongo.deleteOne({ _id: myId });
     return res.status(200).send({ message: "O'chirildi" });
   } catch (error) {
@@ -163,6 +199,34 @@ const refreshingToken = async (req, res) => {
   res.status(200).send({ ...tokens });
 };
 
+const activate = async (req, res) => {
+  try {
+    const argum = await Mongo.findOne({
+      user_activation_link: req.params.link,
+    });
+    if (!argum) {
+      return res.status(400).send({
+        message: "Bundayi topilmad",
+      });
+    }
+
+    if (argum.user_is_active) {
+      return res.status(400).send({ message: "Allaqachon" });
+    }
+
+    argum.user_is_active = true;
+    await argum.save();
+    res.status(200).send({
+      user_is_active: argum.user_is_active,
+      message: "Activated",
+    });
+  } catch (e) {
+    const message = e.message;
+    console.log(message);
+    res.status(404).send({ message: message });
+  }
+};
+
 module.exports = {
   add,
   get,
@@ -171,4 +235,5 @@ module.exports = {
   deleteOne,
   logOutUser,
   refreshingToken,
+  activate,
 };
